use pytest instead of unittest
